feature: Seperate the layout(AppHeader,StickyFooter) and SignInPage
