+++ src/App.js
@@ -1,19 +1,19 @@
 import React from 'react';
-import { useHistory } from 'react-router-dom';
 import './App.css';
-import { StickyFooter, SignIn, AppHeader } from 'designops';
+import { StickyFooter, AppHeader } from 'designops';
 import 'designops/dist/designops.css';
+import RoutingPage from './routingPage';
+import { makeStyles } from '@material-ui/core/styles';
 
 function App() {
-  let history = useHistory();
-  function handleClick() {
-    history.push('/dashboard');
-  }
-  return (
-    <div className="wrapper">
-      <AppHeader />
-      <SignIn firstLabel="Please enter your email" handleClick={handleClick} />
+  const classes = useStyles();
 
+  return (
+    <div className={classes.container}>
+      <AppHeader />
+      <div className={classes.content}>
+        <RoutingPage />
+      </div>
       <StickyFooter
         copyright={`@${new Date().getFullYear()} RXP Services. All rights reserved`}
         footerText="Creative agency making happier humans."
@@ -24,3 +24,15 @@ function App() {
 }
 
 export default App;
+
+const useStyles = makeStyles((theme) => ({
+  container: {
+    height: '100vh',
+    paddingTop: 64,
+    display: 'flex',
+    flexDirection: 'column',
+  },
+  content: {
+    flex: 1,
+  },
+}));
